move 'persistent_upsert' into 'mz_dataflow'
