+++ src/persist/src/operators/mod.rs
@@ -20,7 +20,8 @@ pub mod upsert;
 // https://docs.rs/differential-dataflow/0.12.0/src/differential_dataflow/operators/join.rs.html#487-488
 const DEFAULT_OUTPUTS_PER_YIELD: usize = 1_000_000;
 
-pub(crate) fn split_ok_err<O, E, T, R>(x: (Result<O, E>, T, R)) -> Result<(O, T, R), (E, T, R)> {
+/// lalala
+pub fn split_ok_err<O, E, T, R>(x: (Result<O, E>, T, R)) -> Result<(O, T, R), (E, T, R)> {
     match x {
         (Ok(ok), ts, diff) => Ok((ok, ts, diff)),
         (Err(err), ts, diff) => Err((err, ts, diff)),
